move graph card to global components
