+++ src/components/GraphCard/GraphCard.module.css
@@ -1,9 +1,3 @@
-.expenseGraph {
-    composes: cardComponent from "@styles/style.css";
-    min-height: 480px;
-    
-    grid-column-start: span 2;
-}
 .inner {
     padding: 24px 32px;
     height: 100%;
@@ -38,19 +32,7 @@
     margin-top: 45px;
 }
 
-@media only screen and (max-width: 1440px) {
-    .expenseGraph {
-        order: 1;
-    }
-}
-
 @media only screen and (max-width: 768px) {
-    .expenseGraph {
-        grid-column-start: span 1;
-        min-height: auto;
-        order: 2;
-        height: 330px;
-    }
     .inner{
         padding: 16px;
     }
@@ -58,11 +40,5 @@
         margin-top: 21px;
         min-height: 234px;
         max-height: 234px;
-    }
-}
-
-@media only screen and (max-width: 400px) {
-    .graph {
-        margin-top: 15px;
     }
 }
