+++ src/components/GraphCard/GraphCard.tsx
@@ -1,16 +1,16 @@
 import React, {FC, MouseEvent, ReactNode, useCallback, useState} from "react";
 
 //UI
-import classes from './UserExpenseGraph.module.css';
-import UserExpenseGraph from "./UserExpenseGraph";
+import classes from './GraphCard.module.css';
+import Graph from "./Graph";
 
 //store
 import { useGetExpensesPerLastMonthQuery } from "@store/ExpenseApiSlice/ExpenseApiSlice";
 import { IMonthPickerState } from "@store/UI_store/MonthPickerSlice/MonthPickerInterfaces";
 import { useAppSelector } from "@hooks/storeHooks/useAppStore";
-import UserExpenseGraphLoader from "./UserExpenseGraphLoader";
+import GraphCardLoader from "./GraphCardLoader";
 
-const UserExpenseGraphCard = () => {
+const GraphCard = () => {
     const [loading = true, setLoading] = useState<boolean>();
     const {data: expenses = [], isError: isExpensesError, isLoading: isExpensesLoading, error: Expenses_GET_error } = useGetExpensesPerLastMonthQuery(null);
     const { currentMonth } = useAppSelector<IMonthPickerState>(state => state.MonthPickerSlice);
@@ -26,24 +26,21 @@ const UserExpenseGraphCard = () => {
             }</h3>;
 
     return (
-        <div className={classes.expenseGraph}>
-           {loading ? <UserExpenseGraphLoader/> :
-                <div className={classes.inner}>
-                    <div className={classes.chart__uppernav}>
-                        <div className={classes.chart__titleRange}>
-                            <h2 className={classes.title}>Statistic</h2>
-                            {RangeTitle}
-                        </div>
-                    </div>
-                    <div className={classes.graph}>
-                        <UserExpenseGraph
-                            expenses={expenses}
-                        />
+            loading ? <GraphCardLoader/> :
+            <div className={classes.inner}>
+                <div className={classes.chart__uppernav}>
+                    <div className={classes.chart__titleRange}>
+                        <h2 className={classes.title}>Statistic</h2>
+                        {RangeTitle}
                     </div>
                 </div>
-            }
-        </div>
+                <div className={classes.graph}>
+                    <Graph
+                        expenses={expenses}
+                    />
+                </div>
+            </div>
     )
 }
 
-export default UserExpenseGraphCard
+export default GraphCard
