+++ src/pages/Dashboard/UserGraphCard/UserGraphCard.module.css
@@ -1,0 +1,26 @@
+.UserGraph{
+    composes: cardComponent from "@styles/style.css";
+    min-height: 480px;
+    grid-column-start: span 2;
+}
+
+@media only screen and (max-width: 1440px) {
+    .UserGraph {
+        order: 1;
+    }
+}
+
+@media only screen and (max-width: 768px) {
+    .UserGraph {
+        grid-column-start: span 1;
+        min-height: auto;
+        order: 2;
+        height: 330px;
+    }
+}
+
+@media only screen and (max-width: 400px) {
+    .UserGraph {
+        margin-top: 15px;
+    }
+}
