+++ src/pages/Dashboard/UserGraphCard/UserGraphCard.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+//UI
+import classes from './UserGraphCard.module.css'
+import GraphCard from '@components/GraphCard/GraphCard';
+
+const UserGraphCard = () => {
+    return (
+        <div className={classes.UserGraph}>
+            <GraphCard/>
+        </div>
+    );
+};
+
+export default UserGraphCard;
